♻️ refactor(cli): remove duplicate commit function
- remove `commit_changed_files` function from cli module
- utilize GitOps trait for commit operations

🐛 fix(git_ops): add commit function to GitOps trait

- implement `commit_changed_files` function in GitOps trait
- ensure consistent commit handling across modules

🔧 chore(dependencies): adjust import paths for owo_colors

- move owo_colors imports from cli to git_ops module
